requirements.txt
prometheus-client==0.0.19
speedtest-cli==2.0.2
